refactor: 위젯 Todo 스냅샷 Core 이동
