+++ Application/DevLogData/Sources/Mapper/TodoMapping.swift
@@ -5,6 +5,7 @@
 //  Created by 최윤진 on 2/19/26.
 //
 
+import DevLogCore
 import DevLogDomain
 
 public extension TodoRequest {
